+++ CHANGELOG.md
@@ -332,3 +332,4 @@
 - [2026-04-12 10:26] feat(storage): add async storage stream provider
 - [2026-04-12 11:12] test(navigation): implement widget integration tests for navigation
 - [2026-04-12 11:24] fix(i18n): fix memory leak by cancelling i18n subscription
+- [2026-04-12 11:29] feat(navigation): integrate offline caching for navigation state